% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{ce10.Kundaje.ce10-Excludable}
\alias{ce10.Kundaje.ce10-Excludable}
\title{ce10.Kundaje.ce10-Excludable}
\description{
Defined by Anshul Kundaje, superseded by ce10.Boyle.ce10-Excludable.v2
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
