% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{danRer10.Domingues.Excludableed}
\alias{danRer10.Domingues.Excludableed}
\title{danRer10.Domingues.Excludableed}
\description{
Defined manually using total RNA-seq.
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
