% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{danRer10.Yang.Supplemental_Table_19.ChIP-seq_black_list_in_the_zebrafish_genome}
\alias{danRer10.Yang.Supplemental_Table_19.ChIP-seq_black_list_in_the_zebrafish_genome}
\title{danRer10.Yang.Supplemental_Table_19.ChIP-seq_black_list_in_the_zebrafish_genome}
\description{
Defined via MACS2 peak calling using ChIP-seq (PMID: 33239788)
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
