% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{dm3.Kundaje.dm3-Excludable}
\alias{dm3.Kundaje.dm3-Excludable}
\title{dm3.Kundaje.dm3-Excludable}
\description{
Defined by Anshul Kundaje. Contains heterochromatin chromosomes chr2LHet. Superseded by dm3.Boyle.dm3-Excludable.v2
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
