% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{dm6.Boyle.dm6-Excludable.v2}
\alias{dm6.Boyle.dm6-Excludable.v2}
\title{dm6.Boyle.dm6-Excludable.v2}
\description{
Defined by the Boyle-Lab/Blacklist software, High Signal and Low Mappability
regions
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
