% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{dm6.UCSC.other}
\alias{dm6.UCSC.other}
\title{dm6.UCSC.other}
\description{
Sequence of Ns in the assembly that were not marked as gaps in the AGP (A Golden Path) assembly definition file
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
