% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excluderanges.R
\docType{package}
\name{excluderanges}
\alias{excluderanges}
\title{excluderanges: Genomic coordinates of problematic genomic regions}
\description{
Genomic coordinates of problematic genomic regions that should be
avoided when working with genomic data. GRanges of exclusion
regions (formerly known as blacklisted), Nuclear Mitochondrial Sequences
(NUMTs), centromeres, telomeres, known
heterochromatin regions, etc. (UCSC 'gap' table data). In addition to data
for human (hg19/hg38) and mouse (mm9/mm10) genome assemblies, it provides
excludable regions for the latest T2T and mm39 assemblies. Data for
C. elegans, D. melanogaster, D. rerio, A. thaliana also included.
}
