% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{hg19.Bernstein.Mint_Excludable_hg19}
\alias{hg19.Bernstein.Mint_Excludable_hg19}
\title{hg19.Bernstein.Mint_Excludable_hg19}
\description{
Defined from Mint-ChIP (low input, multiplexed ChIP-seq) data
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
