% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{hg19.Birney.wgEncodeDacMapabilityConsensusExcludable}
\alias{hg19.Birney.wgEncodeDacMapabilityConsensusExcludable}
\title{hg19.Birney.wgEncodeDacMapabilityConsensusExcludable}
\description{
Defined by the ENCODE consortium, includes satellite repeats (CATTC, GAATG,
GAGTG, ACRO1), RepeatMasker repeats (ALR/Alpha, BSR/Beta), centromeric r
epeats, chrM, High/Low mappability islands
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
