% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{hg19.Crawford.wgEncodeDukeMapabilityRegionsExcludable}
\alias{hg19.Crawford.wgEncodeDukeMapabilityRegionsExcludable}
\title{hg19.Crawford.wgEncodeDukeMapabilityRegionsExcludable}
\description{
Defined by the ENCODE consortium, includes satellite repeats (CATTC, GAATG,
GAGTG, ACRO1), RepeatMasker repeats (ALR/Alpha, BSR/Beta), human satellite
repeat HSATII, chrM, ribosomal subunit consensus sequences LSU-rRNA_Hsa,
SSU-rRNA_Hsa. Has extra chromosomes, use keepStandardChromosomes() filtering
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
