% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{hg19.Lareau.hg19_peaks}
\alias{hg19.Lareau.hg19_peaks}
\title{hg19.Lareau.hg19_peaks}
\description{
Regions of high homology to mtDNA (NUMT regions) defined by caleblareau/mitoblacklist
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
