% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{hg19.UCSC.clone}
\alias{hg19.UCSC.clone}
\title{hg19.UCSC.clone}
\description{
Gaps between clones in the same map contig. Has extra chromosomes,
use keepStandardChromosomes() filtering
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
