% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{hg19.UCSC.heterochromatin}
\alias{hg19.UCSC.heterochromatin}
\title{hg19.UCSC.heterochromatin}
\description{
Gaps from large blocks of heterochromatin
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
