% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{hg19.UCSC.scaffold}
\alias{hg19.UCSC.scaffold}
\title{hg19.UCSC.scaffold}
\description{
Gaps between scaffolds in chromosome assemblies. Only non-autosomal chromosomes
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
