% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{hg19.UCSC.short_arm}
\alias{hg19.UCSC.short_arm}
\title{hg19.UCSC.short_arm}
\description{
Gaps on the short arm of the chromosome
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
