% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{hg19.Wold.hg19mitoExcludable}
\alias{hg19.Wold.hg19mitoExcludable}
\title{hg19.Wold.hg19mitoExcludable}
\description{
Definition method unknown
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
