% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{hg19.Yeo.eCLIP_Excludableregions.hg19}
\alias{hg19.Yeo.eCLIP_Excludableregions.hg19}
\title{hg19.Yeo.eCLIP_Excludableregions.hg19}
\description{
Defined from eCLIP data, includes skyscraper, rRNA pseudogene, unreliably
mapped satellite repeat, and low complexity skyscraper peak regions
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
