% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{hg38.Bernstein.Mint_Excludable_GRCh38}
\alias{hg38.Bernstein.Mint_Excludable_GRCh38}
\title{hg38.Bernstein.Mint_Excludable_GRCh38}
\description{
Defined from Mint-ChIP (low input, multiplexed ChIP-seq) data
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
