% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{hg38.Kundaje.GRCh38.Excludable}
\alias{hg38.Kundaje.GRCh38.Excludable}
\title{hg38.Kundaje.GRCh38.Excludable}
\description{
Defined by Anshul Kundaje as a part of ENCODE and modENCODE consortia
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
