% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{hg38.Kundaje.GRCh38_unified_Excludable}
\alias{hg38.Kundaje.GRCh38_unified_Excludable}
\title{hg38.Kundaje.GRCh38_unified_Excludable}
\description{
Defined as a combination of hg38.Lareau.hg38_peaks,
hg38.Boyle.hg38-Excludable.v2, and hg38.Wimberley.peakPass60Perc_sorted,
followed by manual curation. Supersedes hg38.Kundaje.GRCh38.Excludable.
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
