% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{hg38.Lareau.hg38.full.Excludable}
\alias{hg38.Lareau.hg38.full.Excludable}
\title{hg38.Lareau.hg38.full.Excludable}
\description{
ENCODE excludable regions combined with regions of high homology to mtDNA
(NUMT regions)
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
