% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{hg38.Reddy.wgEncodeDacMapabilityConsensusExcludable.hg38}
\alias{hg38.Reddy.wgEncodeDacMapabilityConsensusExcludable.hg38}
\title{hg38.Reddy.wgEncodeDacMapabilityConsensusExcludable.hg38}
\description{
Defined by the ENCODE consortium, includes satellite repeats (CATTC, GAATG,
GAGTG, ACRO1), RepeatMasker repeats (ALR/Alpha, BSR/Beta), centromeric
repeats, chrM, High/Low mappability islands. Has extra chromosomes, use
keepStandardChromosomes() filtering
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
