% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{hg38.UCSC.scaffold}
\alias{hg38.UCSC.scaffold}
\title{hg38.UCSC.scaffold}
\description{
Gaps between scaffolds in chromosome assemblies. Has extra chromosomes,
use keepStandardChromosomes() filtering
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
