% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{hg38.Wimberley.peakPass60Perc_sorted}
\alias{hg38.Wimberley.peakPass60Perc_sorted}
\title{hg38.Wimberley.peakPass60Perc_sorted}
\description{
Defined by the ewimberley/peakPass software
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
