% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{hg38.Wold.hg38mitoExcludable}
\alias{hg38.Wold.hg38mitoExcludable}
\title{hg38.Wold.hg38mitoExcludable}
\description{
Definition method unknown
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
