% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{hg38.Yeo.eCLIP_Excludableregions.hg38liftover.bed.fixed}
\alias{hg38.Yeo.eCLIP_Excludableregions.hg38liftover.bed.fixed}
\title{hg38.Yeo.eCLIP_Excludableregions.hg38liftover.bed.fixed}
\description{
Defined from eCLIP data
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
