% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{mm10.Boyle.mm10-Excludable.v2}
\alias{mm10.Boyle.mm10-Excludable.v2}
\title{mm10.Boyle.mm10-Excludable.v2}
\description{
Defined by the Boyle-Lab/Blacklist software, High Signal and Low Mappability
regions
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
