% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{mm10.Hardison.Excludable.full}
\alias{mm10.Hardison.Excludable.full}
\title{mm10.Hardison.Excludable.full}
\description{
Definition method unknown
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
