% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{mm10.Kundaje.mm10.Excludable}
\alias{mm10.Kundaje.mm10.Excludable}
\title{mm10.Kundaje.mm10.Excludable}
\description{
Defined by Anshul Kundaje as a part of ENCODE and modENCODE consortia
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
