% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{mm10.Lareau.mm10.full.Excludable}
\alias{mm10.Lareau.mm10.full.Excludable}
\title{mm10.Lareau.mm10.full.Excludable}
\description{
ENCODE excludable regions combined with regions of high homology to mtDNA
(NUMT regions)
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
