% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{mm10.UCSC.fragment}
\alias{mm10.UCSC.fragment}
\title{mm10.UCSC.fragment}
\description{
A single gap of 31 bases in chrX_GL456233_random
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
