% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{mm10.UCSC.other}
\alias{mm10.UCSC.other}
\title{mm10.UCSC.other}
\description{
Sequence of Ns in the assembly that were not marked as gaps in the AGP
(A Golden Path) assembly definition file. Has extra chromosomes,
use keepStandardChromosomes() filtering
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
