% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{mm10.UCSC.scaffold}
\alias{mm10.UCSC.scaffold}
\title{mm10.UCSC.scaffold}
\description{
Gaps between scaffolds in chromosome assemblies
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
