% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{mm39.excluderanges}
\alias{mm39.excluderanges}
\title{mm39.excluderanges}
\description{
Defined by the Boyle-Lab/Blacklist software, High Signal and Low Mappability
regions
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
