% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{mm39.UCSC.short_arm}
\alias{mm39.UCSC.short_arm}
\title{mm39.UCSC.short_arm}
\description{
Gaps on the short arm of the chromosome
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
