% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{mm9.Lareau.mm9_peaks}
\alias{mm9.Lareau.mm9_peaks}
\title{mm9.Lareau.mm9_peaks}
\description{
Regions of high homology to mtDNA (NUMT regions) defined by caleblareau/mitoblacklist
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
