% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{mm9.UCSC.contig}
\alias{mm9.UCSC.contig}
\title{mm9.UCSC.contig}
\description{
Gaps between contigs in scaffolds. Has extra chromosomes, use keepStandardChromosomes() filtering
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
