% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{mm9.UCSC.fragment}
\alias{mm9.UCSC.fragment}
\title{mm9.UCSC.fragment}
\description{
Gaps between the contigs of a draft clone. (In this context, a contig is
a set of overlapping sequence reads). Has extra chromosomes,
use keepStandardChromosomes() filtering
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
