% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{mm9.Wold.mm9mitoExcludable}
\alias{mm9.Wold.mm9mitoExcludable}
\title{mm9.Wold.mm9mitoExcludable}
\description{
Definition method unknown
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
