% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{T2T.CHM13.chm13.draft_v1.1.telomere}
\alias{T2T.CHM13.chm13.draft_v1.1.telomere}
\title{T2T.CHM13.chm13.draft_v1.1.telomere}
\description{
Telomere identified by the VGP pipeline (v1.1)
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
