% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{T2T.CHM13.chm13.draft_v2.0.cen_mask}
\alias{T2T.CHM13.chm13.draft_v2.0.cen_mask}
\title{T2T.CHM13.chm13.draft_v2.0.cen_mask}
\description{
Centromeric satellite masking bed file (v2.0)
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
