% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{T2T.Lareau.chm13v2.0_peaks}
\alias{T2T.Lareau.chm13v2.0_peaks}
\title{T2T.Lareau.chm13v2.0_peaks}
\description{
Regions of high homology to mtDNA (NUMT regions) defined by caleblareau/mitoblacklist
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
