% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{T2T.UCSC.censat}
\alias{T2T.UCSC.censat}
\title{T2T.UCSC.censat}
\description{
T2T peri/centromeric satellite annotation (v2.0, 20220329, CHM13 v2.0)
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
