% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{T2T.UCSC.gap}
\alias{T2T.UCSC.gap}
\title{T2T.UCSC.gap}
\description{
Locations of assembly gaps, as determine by strings of 'N' characters (v1.0)
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
