% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{T2T.UCSC.hgUnique.hg38}
\alias{T2T.UCSC.hgUnique.hg38}
\title{T2T.UCSC.hgUnique.hg38}
\description{
Regions unique to the T2T-CHM13 v2.0 assembly compared to the GRCh38/hg38
and GRCh37/hg19 reference assemblies
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
