% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{TAIR10.Klasfeld.arabidopsis_greenscreen_20inputs}
\alias{TAIR10.Klasfeld.arabidopsis_greenscreen_20inputs}
\title{TAIR10.Klasfeld.arabidopsis_greenscreen_20inputs}
\description{
Defined by the green screen pipeline (DOI: 10.1101/2022.02.27.482177)
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
