% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{TAIR10.UCSC.araTha1.gap}
\alias{TAIR10.UCSC.araTha1.gap}
\title{TAIR10.UCSC.araTha1.gap}
\description{
Gaps in the May 2011 Arabidopsis thaliana genome assembly
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
