% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{TAIR10.Wimberley.predicted_excluded_list_sorted_0.6}
\alias{TAIR10.Wimberley.predicted_excluded_list_sorted_0.6}
\title{TAIR10.Wimberley.predicted_excluded_list_sorted_0.6}
\description{
Defined by the ewimberley/peakPass software
}
\author{
Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com}
}
