% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excluderanges.R
\docType{package}
\name{excluderanges}
\alias{excluderanges}
\title{excluderanges: Genomic coordinates of problematic genomic regions}
\description{
Genomic coordinates of problematic genomic regions that should be
avoided when working with genomic data. GRanges of exclusion
regions (formerly known as blacklisted), centromeres, telomeres, known
heterochromatin regions, etc. (UCSC 'gap' table data). Primarily for human
and mouse genomes, hg19/hg38 and mm9/mm10 genome assemblies.
}
